rename CosmosGraphDB binding component to CosmosDBGremlinAPI #1246 (#1276)
